added GNS message config for u-blox, fixed task runtime problem

+++ FreeRTOS/Source/include/task.h
@@ -69,8 +69,8 @@ extern "C" {
 struct tskTaskControlBlock; /* The old naming convention is used to prevent breaking kernel aware debuggers. */
 typedef struct tskTaskControlBlock* TaskHandle_t;
 /*MOD*/
-uint32_t taskGetRunTimeCount(TaskHandle_t task);
-void taskSetRunTimeCount(TaskHandle_t task, uint32_t cnt);
+uint64_t taskGetRunTimeCount(TaskHandle_t task);
+void taskSetRunTimeCount(TaskHandle_t task, uint64_t cnt);
 void taskGetPriorities(TaskHandle_t task, uint32_t* basePriority, uint32_t* currentPriority);
 
 
@@ -147,7 +147,7 @@ typedef struct xTASK_STATUS
 	eTaskState eCurrentState;		/* The state in which the task existed when the structure was populated. */
 	UBaseType_t uxCurrentPriority;	/* The priority at which the task was running (may be inherited) when the structure was populated. */
 	UBaseType_t uxBasePriority;		/* The priority to which the task will return if the task's current priority has been inherited to avoid unbounded priority inversion when obtaining a mutex.  Only valid if configUSE_MUTEXES is defined as 1 in FreeRTOSConfig.h. */
-	uint32_t ulRunTimeCounter;		/* The total run time allocated to the task so far, as defined by the run time stats clock.  See http://www.freertos.org/rtos-run-time-stats.html.  Only valid when configGENERATE_RUN_TIME_STATS is defined as 1 in FreeRTOSConfig.h. */
+	uint64_t ulRunTimeCounter;		/* The total run time allocated to the task so far, as defined by the run time stats clock.  See http://www.freertos.org/rtos-run-time-stats.html.  Only valid when configGENERATE_RUN_TIME_STATS is defined as 1 in FreeRTOSConfig.h. */
 	StackType_t *pxStackBase;		/* Points to the lowest address of the task's stack area. */
 	configSTACK_DEPTH_TYPE usStackHighWaterMark;	/* The minimum amount of stack space that has remained for the task since the task was created.  The closer this value is to zero the closer the task has come to overflowing its stack. */
 } TaskStatus_t;
